pyproject.toml
[tool.poetry]
name = "lms-project-drf"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "lms_project_drf"}]

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.2.4"
djangorestframework = "^3.14.0"


[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.7"
pillow = "^10.0.0"
ipython = "^8.14.0"
django-filter = "^23.2"
djangorestframework-simplejwt = "^5.2.2"
coverage = "^7.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

